바로 삭제 보다는 한 번더 확인차 묻고 삭제할 수 있도록 수정

+++ src/main/resources/static/js/article.js
@@ -1,15 +1,32 @@
 // 삭제
 const deleteButton = document.getElementById('delete-btn');
 
+// if (deleteButton) {
+//     deleteButton.addEventListener('click', event => {
+//         let id = document.getElementById('article-delete-id').value;
+//         fetch(`/api/articles/${id}`, { // BlogApiController의 DeleteMapping이 실행된다.
+//             method: 'DELETE'
+//         })
+//             .then(() => {
+//                 alert("삭제 완료 되었습니다.");
+//                 location.replace('/articles');
+//             })
+//     });
+// }
 if (deleteButton) {
     deleteButton.addEventListener('click', event => {
-        let id = document.getElementById('article-delete-id').value;
-        fetch(`/api/articles/${id}`, { // BlogApiController의 DeleteMapping이 실행된다.
-            method: 'DELETE'
-        })
-            .then(() => {
-                alert("삭제 완료 되었습니다.");
-                location.replace('/articles');
+
+        const confirmDeletion = confirm("정말 삭제하시겠습니까?");
+
+        if (confirmDeletion) {
+            let id = document.getElementById('article-delete-id').value;
+            fetch(`/api/articles/${id}`, { // BlogApiController의 DeleteMapping이 실행된다.
+                method: 'DELETE'
             })
+                .then(() => {
+                    alert("삭제가 완료 되었습니다.");
+                    location.replace('/articles');
+                })
+        } 
     });
 }
